automatic version based on git tag
Update makefile to set current version based on information in git. If
the current commit has a signed tag "v1.2.3" the current version will be
set to "1.2.3". If there is no tag for the current a version string
based on the latest tag as per `git describe` will be used.

Also sets a build date for good measure