Add support for Masked Environment variables
The purpose of these is to be able to pass in for example credentials to
your docker container at runtime, but not have them exposed through a
GET with the Eremetic API.

They are not encrypted, and will still be viewable if someone has access
to the machine running the task, as they are merged to the env field
before being passed to Mesos.

See the readme for more information, and follow best practices for
secure communication for REST-ful APIs.